Add keyboard functionality and refactor functions
Creating keyboard functionality by adding an eventListener to the
window object and matching the keydown event to the data-keycode
attribute on each button before simulating a click on a that button.

Refactoring the creation of individual operation button event listeners
into a single for loop using the id of the button.